[Det] EfficientDet Wheat-Detection/requirements.txt
bottle~=0.12.18
sklearn~=0.0
joblib~=0.14.1
wheel
pandas~=1.0.3
requests~=2.23.0
scipy~=1.4.1
numpy~=1.18.4
opencv-python~=4.5.1.48
torchvision~=0.9.1
timm~=0.4.5
scikit-learn~=0.23.0
pydicom~=2.1.2
tqdm~=4.60.0
pillow~=8.2.0
scikit-image~=0.18.1
datetime~=4.3
albumentations~=0.5.2
effdet


